+++ controllers/invoice_controller.py
@@ -4,7 +4,7 @@ from dtos.invoice_models import InvoiceResponse, CreateInvoiceRequest, UpdateInv
 from helper.role_helper import RoleHelper
 from helper.api_helper import APIHelper
 from models.invoice import Invoice
-from models.client import Client
+from models.user import User
 from datetime import date
 from config.db_config import SessionLocal
 from fastapi import HTTPException
@@ -13,9 +13,14 @@ import i18n
 class InvoiceController:
     
     def get_all_invoice(user:UserModel)->BaseResponseModel:
-        RoleHelper.require_role(["lawyer","admin"],user)
+        RoleHelper.require_role(["lawyer","admin","client"],user)
         with SessionLocal() as db:
-            invoices=db.query(Invoice).all()
+            if user.role=="client":
+                invoices=db.query(Invoice).filter(Invoice.client_id==user.id).all()
+            elif user.role == "admin":
+                invoices=db.query(Invoice).all()
+            else:
+                invoices=db.query(Invoice).filter(Invoice.created_by==user.id).all()
             
             if not invoices:
                 raise HTTPException(
@@ -33,8 +38,10 @@ class InvoiceController:
     def get_invoice_by_id(invoice_id:int,user:UserModel)->BaseResponseModel:
         RoleHelper.require_role(["lawyer","admin"],user)
         with SessionLocal() as db:
-            
-            invoice = db.query(Invoice).filter(Invoice.id==invoice_id).first()
+            if user.role !="admin":
+                invoice = db.query(Invoice).filter(Invoice.id==invoice_id,Invoice.created_by==user.id).first()
+            else:
+                invoice = db.query(Invoice).filter(Invoice.id==invoice_id).first()
             
             if not invoice:
                 raise HTTPException(status_code=404, detail=i18n.t("translations.INVOICE_NOT_FOUND"))
@@ -47,12 +54,13 @@ class InvoiceController:
     def create_invoice(invoice_data:CreateInvoiceRequest,user:UserModel)->BaseResponseModel:
         RoleHelper.require_role(["lawyer","admin"],user)
         with SessionLocal() as db:
-            client=db.query(Client).filter(Client.id==invoice_data.client_id,Client.is_deleted==False).first()
+            client=db.query(User).filter(User.id==invoice_data.client_id,User.is_deleted==False, User.role=="client").first()
             
             if not client:
                 raise HTTPException(status_code=404, detail=i18n.t("translations.CLIENT_NOT_FOUND"))
             
             invoice=db.query(Invoice).filter(Invoice.invoice_number==invoice_data.invoice_number).first()
+            
             if invoice:
                 raise HTTPException(status_code=400, detail=i18n.t("translations.INVOICE_NUM_EXISTS")
 )
@@ -84,8 +92,13 @@ class InvoiceController:
             if not invoice:
                 raise HTTPException(status_code=404, detail=i18n.t("translations.INVOICE_NOT_FOUND"))
             
+            if invoice.created_by!=user.id and user.role=="lawyer":
+                raise HTTPException(
+                    status_code=403, detail=i18n.t("translations.UNAUTHORIZED")
+                )
+            
             if update_data.client_id is not None:
-                new_client = db.query(Client).filter(Client.id == update_data.client_id, Client.is_deleted == False).first()
+                new_client = db.query(User).filter(User.id == update_data.client_id, User.is_deleted == False, User.role=="client").first()
                 if not new_client:
                     raise HTTPException(
                         status_code=404, detail=i18n.t("translations.CLIENT_NOT_FOUND")
@@ -112,6 +125,11 @@ class InvoiceController:
             
             if not invoice:
                 raise HTTPException(status_code=404, detail=i18n.t("translations.INVOICE_NOT_FOUND"))
+            
+            if invoice.created_by!=user.id and user.role=="lawyer":
+                raise HTTPException(
+                    status_code=403, detail=i18n.t("translations.UNAUTHORIZED")
+                )
             
             db.delete(invoice)
             db.commit()
